Ditch resources.py
