--- integration-testing/rnode_testing/resources.py
@@ -1,20 +1,0 @@
-import os
-
-
-def get_absolute_path_for_mounting(relative_path, mount_dir=None):
-    """Drone runs each job in a new Docker container FOO. That Docker
-    container has a new filesystem. Anything in that container can read
-    anything in that filesystem. To read files from HOST, it has to be shared
-    though, so let's share /tmp:/tmp. You also want to start new Docker
-    containers, so you share /var/run/docker.sock:/var/run/docker.sock. When
-    you start a new Docker container from FOO, it's not in any way nested. You
-    just contact the Docker daemon running on HOST via the shared docker.sock.
-    So when you start a new image from FOO, the HOST creates a new Docker
-    container BAR with brand new filesystem. So if you tell Docker from FOO to
-    mount /MOUNT_DIR:/MOUNT_DIR from FOO to BAR, the Docker daemon will actually mount
-    /MOUNT_DIR from HOST to BAR, and not from FOO to BAR.
-    """
-
-    if mount_dir is not None:
-        return os.path.join(mount_dir, relative_path)
-    return os.path.abspath(os.path.join('resources', relative_path))
